+++ .changeset/ten-wasps-serve.md
@@ -1,0 +1,8 @@
+---
+'@backstage/plugin-catalog-backend-module-github': patch
+'@backstage/core-components': patch
+'@backstage/plugin-search-react': patch
+'@backstage/plugin-playlist': patch
+---
+
+Internal restructure to avoid circular imports
