+++ packages/core-components/src/layout/Sidebar/MobileSidebar.tsx
@@ -25,21 +25,12 @@ import Typography from '@material-ui/core/Typography';
 import CloseIcon from '@material-ui/icons/Close';
 import MenuIcon from '@material-ui/icons/Menu';
 import { orderBy } from 'lodash';
-import React, { createContext, useEffect, useState, useContext } from 'react';
+import React, { useEffect, useState, useContext } from 'react';
 import { useLocation } from 'react-router-dom';
 import { SidebarOpenStateProvider } from './SidebarOpenStateContext';
 import { SidebarGroup } from './SidebarGroup';
 import { SidebarConfigContext, SidebarConfig } from './config';
-
-/**
- * Type of `MobileSidebarContext`
- *
- * @internal
- */
-export type MobileSidebarContextType = {
-  selectedMenuItemIndex: number;
-  setSelectedMenuItemIndex: React.Dispatch<React.SetStateAction<number>>;
-};
+import { MobileSidebarContext } from './MobileSidebarContext';
 
 /**
  * Props of MobileSidebar
@@ -148,16 +139,6 @@ const OverlayMenu = ({
     </Drawer>
   );
 };
-
-/**
- * Context on which `SidebarGroup` is currently selected
- *
- * @internal
- */
-export const MobileSidebarContext = createContext<MobileSidebarContextType>({
-  selectedMenuItemIndex: -1,
-  setSelectedMenuItemIndex: () => {},
-});
 
 /**
  * A navigation component for mobile screens, which sticks to the bottom.
