+++ packages/core-components/src/layout/Sidebar/MobileSidebarContext.ts
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2023 The Backstage Authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import React, { createContext } from 'react';
+
+/**
+ * Type of `MobileSidebarContext`
+ *
+ * @internal
+ */
+export type MobileSidebarContextType = {
+  selectedMenuItemIndex: number;
+  setSelectedMenuItemIndex: React.Dispatch<React.SetStateAction<number>>;
+};
+
+/**
+ * Context on which `SidebarGroup` is currently selected
+ *
+ * @internal
+ */
+export const MobileSidebarContext = createContext<MobileSidebarContextType>({
+  selectedMenuItemIndex: -1,
+  setSelectedMenuItemIndex: () => {},
+});
