+++ packages/core-components/src/layout/Sidebar/SidebarGroup.tsx
@@ -24,7 +24,7 @@ import React, { useContext } from 'react';
 import { useLocation } from 'react-router-dom';
 import { Link } from '../../components/Link/Link';
 import { SidebarConfig, SidebarConfigContext } from './config';
-import { MobileSidebarContext } from './MobileSidebar';
+import { MobileSidebarContext } from './MobileSidebarContext';
 import { useSidebarPinState } from './SidebarPinStateContext';
 
 /**
