+++ plugins/catalog-backend-module-github/src/lib/annotation.ts
@@ -16,10 +16,10 @@
 
 /**
  * The value of this annotation is the so-called login that identifies a user on
-[GitHub](https://github.com) (either the public one, or a private GitHub
-Enterprise installation) that is related to this entity. It is on the format
-`<username>`, and is the same as can be seen in the URL location bar of the
-browser when viewing that user.
+ * [GitHub](https://github.com) (either the public one, or a private GitHub
+ * Enterprise installation) that is related to this entity. It is on the format
+ * `<username>`, and is the same as can be seen in the URL location bar of the
+ * browser when viewing that user.
  *
  * @public
  */
@@ -27,10 +27,10 @@ export const ANNOTATION_GITHUB_USER_LOGIN = 'github.com/user-login';
 
 /**
  * The value of this annotation is the so-called slug that identifies a team on
-[GitHub](https://github.com) (either the public one, or a private GitHub
-Enterprise installation) that is related to this entity. It is on the format
-`<organization>/<team>`, and is the same as can be seen in the URL location bar
-of the browser when viewing that team.
+ * [GitHub](https://github.com) (either the public one, or a private GitHub
+ * Enterprise installation) that is related to this entity. It is on the format
+ * `<organization>/<team>`, and is the same as can be seen in the URL location
+ * bar of the browser when viewing that team.
  *
  * @public
  */
